# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Health API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/foods": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Foods",
        "operationId": "foods_api_foods_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[BCDEFGHKMNPQRSTUVWXY]$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          },
          {
            "name": "source_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Name"
            }
          },
          {
            "name": "preparation_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Preparation State"
            }
          },
          {
            "name": "min_protein",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Protein"
            }
          },
          {
            "name": "min_protein_density",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Protein Density"
            }
          },
          {
            "name": "min_fiber",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Fiber"
            }
          },
          {
            "name": "max_energy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Energy"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "relevance",
                "name",
                "name_desc",
                "group_asc",
                "group_desc",
                "code_asc",
                "code_desc",
                "source_asc",
                "source_desc",
                "energy_asc",
                "energy_desc",
                "protein_asc",
                "protein_desc",
                "protein_density_asc",
                "protein_density_desc",
                "fat_asc",
                "fat_desc",
                "carbs_asc",
                "carbs_desc",
                "fiber_asc",
                "fiber_desc"
              ],
              "type": "string",
              "default": "relevance",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Create Food",
        "description": "Register a branded food with its label nutrition and portions.\n\nValidates the barcode check digit, \"davon\" rows, kJ against kcal, and kcal\nagainst the macros. Use `dry_run=true` first; warnings do not block writing.",
        "operationId": "create_food_api_foods_post",
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dry Run"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodRegistration"
                }
              }
            }
          },
          "409": {
            "description": "Barcode already registered; `slug` names the food"
          },
          "422": {
            "description": "Invalid or self-contradicting values; nothing written"
          }
        }
      }
    },
    "/api/foods/facets": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Facets",
        "description": "Filter values with counts for the same filters as `GET /api/foods`.\n\nEach dimension is counted under the other filters, so the counts show what\nselecting a value would yield.",
        "operationId": "facets_api_foods_facets_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[BCDEFGHKMNPQRSTUVWXY]$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          },
          {
            "name": "preparation_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Preparation State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogFacets"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/foods/barcode/{barcode}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Food By Barcode",
        "description": "Exact lookup of a scanned or photographed EAN/UPC barcode.",
        "operationId": "food_by_barcode_api_foods_barcode__barcode__get",
        "parameters": [
          {
            "name": "barcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Barcode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/foods/{slug}/sources/{source_id}": {
      "patch": {
        "tags": [
          "catalog"
        ],
        "summary": "Patch Source Text",
        "description": "Add a label's legal name or ingredients read from a later photo.\n\nOnly text fields; a label with different nutrition is a new source. BLS\nsources cannot be edited.",
        "operationId": "patch_source_text_api_foods__slug__sources__source_id__patch",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceTextUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodDetail"
                }
              }
            }
          },
          "404": {
            "description": "No such food or source"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/foods/{slug}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Food Detail",
        "operationId": "food_detail_api_foods__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CatalogFacets": {
        "properties": {
          "foods": {
            "type": "integer",
            "title": "Foods"
          },
          "any_group": {
            "type": "integer",
            "title": "Any Group"
          },
          "sources": {
            "type": "integer",
            "title": "Sources"
          },
          "kinds": {
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            },
            "type": "array",
            "title": "Kinds"
          },
          "source_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Names"
          },
          "preparation_states": {
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            },
            "type": "array",
            "title": "Preparation States"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/FacetGroup"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "foods",
          "any_group",
          "sources",
          "kinds",
          "source_names",
          "preparation_states",
          "groups"
        ],
        "title": "CatalogFacets",
        "description": "Filter values with counts. Each list is counted under every active\nfilter except its own, and holds the values with at least one food plus\nthe selected value (with count 0 if nothing matches)."
      },
...
      "FacetGroup": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "count"
        ],
        "title": "FacetGroup"
      },
      "FacetValue": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "value",
          "count"
        ],
        "title": "FacetValue"
      },
      "FoodDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "preparation_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preparation State"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceDetail"
            },
            "type": "array",
            "title": "Sources"
          },
          "portions": {
            "items": {
              "$ref": "#/components/schemas/Portion"
            },
            "type": "array",
            "title": "Portions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "aliases",
          "kind",
          "preparation_state",
          "brand",
          "barcode",
          "sources",
          "portions"
        ],
        "title": "FoodDetail"
      },
      "FoodInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 14
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "aliases": {
            "items": {
              "type": "string",
              "maxLength": 200,
              "minLength": 1
            },
            "type": "array",
            "title": "Aliases",
            "default": []
          },
          "source_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source Name",
            "default": "Product label"
          },
          "source_food_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Food Name"
          },
          "ingredients_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients Text"
          },
          "nutrition": {
            "$ref": "#/components/schemas/NutritionInput"
          },
          "portions": {
            "items": {
              "$ref": "#/components/schemas/PortionInput"
            },
            "type": "array",
            "title": "Portions",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "nutrition"
        ],
        "title": "FoodInput",
        "description": "A branded food from a label photo or product database."
      },
      "FoodPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FoodSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "FoodPage"
      },
      "FoodRegistration": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "type": "array",
            "title": "Warnings"
          },
          "food": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FoodDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "dry_run",
          "slug",
          "warnings",
          "food"
        ],
        "title": "FoodRegistration"
      },
      "FoodSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "preparation_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preparation State"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NutritionSource"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "aliases",
          "kind",
          "preparation_state",
          "brand",
          "barcode",
          "source_count",
          "source"
        ],
        "title": "FoodSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Issue": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "Issue"
      },
...
      "Portion": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "quantity": {
            "type": "string",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "quantity",
          "unit"
        ],
        "title": "Portion"
      },
      "PortionInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "package",
              "serving",
              "piece",
              "household"
            ],
            "title": "Kind"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "enum": [
              "g",
              "ml"
            ],
            "title": "Unit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "kind",
          "quantity",
          "unit"
        ],
        "title": "PortionInput"
      },
...
      "SourceTextUpdate": {
        "properties": {
          "food_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Name"
          },
          "ingredients_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SourceTextUpdate",
        "description": "Text read later from the same label, e.g. a second photo of a round cup."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}